// generated/src/app/api-models/al-lob-letter-events.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AlLobLetterEvents {
  lineOfBusiness: string;
  letterEventId: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  replaceFlag: string;
}